+++ cmd/dasel/main.go
@@ -1,14 +1,21 @@
 package main
 
 import (
-	"github.com/tomwright/dasel/v2/internal/command"
+	"io"
 	"os"
+
+	"github.com/tomwright/dasel/v3/internal/cli"
+	_ "github.com/tomwright/dasel/v3/parsing/d"
+	_ "github.com/tomwright/dasel/v3/parsing/json"
+	_ "github.com/tomwright/dasel/v3/parsing/toml"
+	_ "github.com/tomwright/dasel/v3/parsing/yaml"
 )
 
 func main() {
-	cmd := command.NewRootCMD()
-	if err := cmd.Execute(); err != nil {
-		cmd.PrintErrln("Error:", err.Error())
-		os.Exit(1)
+	var stdin io.Reader = os.Stdin
+	fi, err := os.Stdin.Stat()
+	if err != nil || (fi.Mode()&os.ModeNamedPipe == 0) {
+		stdin = nil
 	}
+	cli.MustRun(stdin, os.Stdout, os.Stderr)
 }